refactor: relocate window controllers
