+++ automl.py
@@ -16,8 +16,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-"""Project version information."""
+"""Compatibility wrapper exposing the AutoML launcher as ``automl``.
 
-VERSION = "0.2.104"
+This module re-exports all public objects from :mod:`AutoML` so tests and
+user scripts can ``import automl`` regardless of filename casing.
+"""
 
-__all__ = ["VERSION"]
+from AutoML import *  # noqa: F401,F403
+
+__all__ = [name for name in globals() if not name.startswith("_")]
+
