+++ mainappsrc/__init__.py
@@ -28,7 +28,6 @@ _submodule_map = {
     # Core modules (dependencies first)
     "page_diagram": "core.page_diagram",
     "event_dispatcher": "core.event_dispatcher",
-    "window_controllers": "core.window_controllers",
     "diagram_renderer": "core.diagram_renderer",
     # UI helpers migrated from core
     "app_lifecycle_ui": "ui.app_lifecycle_ui",
